add undo/redo functionality (#71)
Implemented POST `/mixtape/{public_id}/undo` and POST `/mixtape/{public_id}/redo` endpoints in the mixtape router. These endpoints use the version history in the mixtape_snapshot database tabelsto implement "undo" and "redo" the same way they work in most applications: in other words, they hold a single stack of undoable/redoable actions (no need to worry about branching— After an undo, if the user makes a new edit (PUT), we can break the old redo chain). They will then return the MixtapeResponse object.

In order for these to work cleanly, we made database changes to represent the undo/redo stack. Recall the DB model design: mixtape_snapshot is essentially version history (an audit log) for the mixtape table, capturing a snapshot of each entry in the mixtape table as it has existed at every moment it gets updated throughout history, including the current values, with the exception of immutable columns like id and public_id. This means that the current information is always duplicated in both tables (which is a bit wasteful), but provides full audit trails for version history. This means that for a given mixtape entry, there should be at least one mixtape_snapshot entry (the only case where it would have exactly one is if was created once and never modified after that).

So with this in mind, the way we chose to add this new behavior is to add three new columns to both mixtape and mixtape_snapshot: undo_to_version, resembles_version, and redo_to_version. Semantics:
- undo_to_version = the version you would use as the target_snapshot if you undid from this version.
- resembles_version = the version that this new version resembles.
- redo_to_version = the version you would use as the target_snapshot if you redid from this version.
This three-pointer scheme works because:
- Every node in mixtape_snapshot is immutable and append-only — when you undo/redo, you don’t "move" a pointer, you create a new node and wire up the links between it and its immediate undo/redo neighbors.
- At any point, undo = follow undo_to_version, redo = follow redo_to_version.
- Multiple undos push backward along undo_to_version links if they exist.
- Multiple redos push forward along redo_to_version links if they exist.
- An edit (normal PUT request after an undo or a redo) breaks the chain and sets redo_to_version of the new version to NULL.
- Note that the snapshot table remains immutable and append-only, so even when we "follow" links we are just using them as references to generate a new (duplicate) snapshot, and pointing back to that duplicate via the resembles_version.

This is essentially a doubly-linked list of all possible current-head positions, stored in the audit table.

So let's consider an example case: the mixtape is at version 5 (because it started at version 1 with a POST request and the user then made four subsequent PUT requests). Let's say the user then triggers an undo. This will generate a new snapshot with a version of 6, but with an undo pointer which would point to 3 and a redo pointer which would point to 5, because it resembles version 4. In case that doesn't make sense yet, I will perform one more step to make it clearer. In fact, let's consider all three of the next possible cases:
(1) if the user then triggers an undo again, it will insert a new snapshot with a version of 7, but with an undo pointer which would point to 2 and a redo pointer which would point to 4, and a resembles version of 3.
(2) However the user then triggers a redo, it will insert a new snapshot with a version of 7, but with an undo pointer which would point to 4 and a redo pointer which would point to None, and a resembles version of 5.
(3) Alternatively, if the user just does a normal PUT request, it would break the chain and insert a new snapshot with a version of 7, with an undo pointer which would point to 6, a redo pointer of None and a resembles version of None.

Note also that mixtape_snapshot_track is the corollary of mixtape_snapshot for mixtape_tracks. mixtape_snapshot_tracks are just linked to mixtape_snapshot ids, so we can just figure out the tracks that a mixtape had at version x by finding the mixtape_snapshot by version x. When we undo/redo a mixtape, the mixtape tracks get fetched from the snapshot along with it.

At the API level, we should add fields "can_redo", "can_undo", and "version" in the MixtapeResponse object, so the client knows if they can undo/redo as well as the current version. The logic is that they can undo if undo_to_version is non-null, and they can redo if redo_to_version is non-null. If clients ignore that field and try to undo a non-undoable mixtape, they'll get a 400 bad request. Same if they try to redo a non-redoable mixtape.

Note that users may submit multiple undos in quick sequence, and we should make sure to handle them correctly, so we should select for update like we do in other components.