organize components throughout the app appropriately (#89)
Previously they were just a laundry list in a single directory. Now, they are placed in relevant locations throughout the app.

Also, the `@` import has been improved so you no longer need to call `@/app` every time, just `@`. This also makes it consistent with jest, which didn't require/expect the `/app` part of the path since it was already adding it automatically.

Finally, component tests have been placed in the same directory as the component. this makes it easier to see when components are untested.